Initial commit
This is the start of a small command line tool, spec2json. The main goal
for it is supporting SerenityOS LibJS/LibWeb development by providing
spec section metadata and algorithm steps in an easy-to-use format, i.e.
JSON. Actual diffing of spec comments in the source code with the
extracted data will be done separately, so this project isn't directly
tied and limited to those goals.
Supported formats include Ecmarkup (already implemented), Bikeshed, and
Wattsi (both stubbed), which should cover the majority of web specs.